requirements-dev.txt
# Development and build dependencies
nuitka
ordered-set
zstandard
flet-desktop==0.28.3
